requirements.txt
gradio
transformers
pillow